Start work on thin subs
A thin sub is a sub that only holds sub commands, it does not have its
own main sub script, but relies instead on a central one that provides
all the functionality for it.

This feature allows the user to keep the sub software up to date while
their own scripts evolve separately.